# spec/api.yaml
...
openapi: 3.0.3
info:
  title: TUD Assistance Backbone API
  description: Component for analyzing learning process data in the form of xAPI statements and providing feedback to
    learners' interactions and assistance and suggestions with regard to the user's learning state and the corresponding
    learning content.
  license:
    name: GNU General Public License
    url: https://www.gnu.org/licenses/old-licenses/gpl-3.0.de.html
  version: 1.0.0
...
security:
- bearerAuth: []
paths:
  /api/v1/administration/setting/{key}:
    get:
      tags:
      - Administration
      summary: Get setting
      description: Get the setting with the specified key.
      operationId: getSetting
      parameters:
      - $ref: '#/components/parameters/settingKey'
      responses:
        200:
          description: Successfully retrieved setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setting'
        400:
          description: Bad request
        401:
          description: Authentication failed
        404:
          description: A setting with the specified key does not exist
    put:
      tags:
      - Administration
      summary: Update setting
      description: Update the setting with the specified key.
      operationId: updateSetting
      parameters:
      - $ref: '#/components/parameters/settingKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Setting'
      responses:
        204:
          description: Successfully updated setting
        400:
          description: Bad request
        401:
          description: Authentication failed
        422:
          description: Unprocessable entity
      x-codegen-request-body-name: body
...
components:
  schemas:
...
    Setting:
      type: object
      description: Schema for representing a setting of the system.
      required:
      - key
      properties:
        key:
          type: string
          description: The unique key of the setting.
          example: assistance_stopped
        value:
          description: The value of the setting.
          example: true
...
  parameters:
...
    settingKey:
      name: key
      in: path
      description: Unique key of a setting.
      required: true
      schema:
        type: string
        example: assistance_stopped
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT